README.md
William E Bales spent 20 years in the scientific design and creation of technical processes that allows the evolutionary VPOS Quantum Optical Binary to undercode any data regardless of operating system or programming language and secure its data with 100% accuracy and deliver it from point A to point B with ZERO hacking ability, no virus can attach to any file or be embedded after static rendering, trojans are automatically listed with the vendor as a non executable until they use corrective action, There is no pirating ability by any third party or device authorized in the world. You cannot write code around this technology, you cannot write code to alter or copy this technology. This is an "optical code" that doesn't use numerical symbols for coding by a standard keyboard. Meaning, the next CPU that is going to evolutionize the computer science is 100% "optical processing". VPOS Technologies Corporation is the only CPU architecturally designed for the speed of light and completely changes the Northbridge, RAM, and circuitry into the new world of optical processing. Welcome to the final frontier. @VPOSTechnology @WilliamEBales